Fix writeUpdateStats is now double
The timeSpent gets assigned as a double, so it should be a double in the
struct as well.

+++ tests/suite/codec/timed/encoderStats.h
@@ -13,7 +13,7 @@ namespace suite {
   struct writeUpdateStats
   {
     uint timeRequired;
-    uint timeSpent;
+    double timeSpent;
 
     // Was the frame encoded within the time budget?
     bool encodedInTime() const { return timeSpent <= timeRequired; }
